Added function compatibility to C
- Added function for conversion of strings from int and float.
- Added function for strcat
- Changed CType from double to float
-Fixed nodetype for binaryExpr

+++ src/nodetype/PrimitiveNodeType.java
@@ -312,7 +312,7 @@ public enum PrimitiveNodeType implements NodeType {
             case STRING:
                 return "char *";
             case REAL:
-                return "double";
+                return "float";
             case BOOL:
                 return "bool";
             case INT:
